refactor : 고객 패키지 분리
